requirements.txt
aiosmtplib==3.0.2
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
asyncmy==0.2.10
bcrypt==4.2.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
decouple==0.0.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-mail==1.4.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jwt==1.3.1
kombu==5.4.2
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysqlclient==2.2.6
passlib==1.7.4
prompt_toolkit==3.0.48
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
